Refactored store files and updated related imports
