+++ packages/common/store/actions/index.js
@@ -4,8 +4,8 @@ export * from '@bookmark/store/action';
 export * from '@network/store/action';
 export * from '@settings/store/actions';
 export * from '@transaction/store/actions';
-export * from '@dpos/store/actions/voting';
-export * from '@dpos/store/actions/watchList';
+export * from '@dpos/validator/store/actions/voting';
+export * from '@dpos/validator/store/actions/watchList';
 export * from '@updater/store/actions/appUpdates';
 export * from './loading';
 export * from './service';
