+++ packages/common/store/reducers/index.js
@@ -5,7 +5,7 @@ export { default as network } from '@network/store/reducer';
 export { default as settings } from '@settings/store/reducer';
 export { default as transactions } from '@transaction/store/reducer';
 export { default as appUpdates } from '@updater/store/reducers/appUpdates';
-export { default as voting } from '@dpos/store/reducers/voting';
-export { default as watchList } from '@dpos/store/reducers/watchList';
+export { default as voting } from '@dpos/validator/store/reducers/voting';
+export { default as watchList } from '@dpos/validator/store/reducers/watchList';
 export { default as service } from './service';
 export { default as loading } from './loading';
