+++ src/modules/dpos/validator/store/actions/voting.js
@@ -3,9 +3,9 @@ import { tokenMap } from '@token/configuration/tokens';
 import { MODULE_ASSETS_NAME_ID_MAP } from '@transaction/configuration/moduleAssets';
 import { create } from '@transaction/utilities/api';
 import { getAccount } from '@wallet/utilities/api';
-import { getVotes } from '@dpos/utilities/api';
 import { isEmpty } from '@common/utilities/helpers';
 import { timerReset } from '@wallet/store/action';
+import { getVotes } from '../../utils/api';
 import actionTypes from './actionTypes';
 
 export const votesReset = () => ({
