+++ src/modules/dpos/validator/store/middlewares/voting.js
