+++ src/modules/dpos/validator/store/middlewares/watchList.js
