+++ src/modules/dpos/validator/store/reducers/voting.js
