+++ src/modules/dpos/validator/store/reducers/watchList.js
